test: написать тесты для /history, /prioritized
